refactored codebase to authcontroller and usercontroller separately
